uitext: Add a "go back" option to the "proceed" prompt
When one mistakenly skips the last item of the "selection" prompt this
option starts again at the last item whereas the existing option "do
again" reprocesses the whole list which can be long.

Add an accumulator to the function interactAndPropagateChanges() that
passes it as a newly added accumulator of the function interact() which
simply passes it as the accumulator of its subfuntion loop().  In
interactAndPropagateChanges() additionally append the accumulator to the
main list when we want to go back to the selection.