Util: make splitIntoWords() parse an escape character
splitIntoWords() is called to construct the shell commands (processing
the arguments of the parameters 'rshargs' and 'sshargs') and to process
the filename of an 'include' directive.

The main advantage of the change is that it enables to specify proxy
commands (that contain spaces) as arguments to ssh.

The escape mechanism is not that of C: any character following the
backslash is taken as is.