Uitext.selectAction: print error if exception raised while reading stdin
This change prints an error message in the case (amongst others) of the
rare condition that Unison can write to the terminal but fails to read
from it.

The exceptions are raised only inside Uitext.getInput which is called
only by Uitext.selectAction.  So catch the exceptions there to print a
message then re-raise them (for now).