Fix documentation: merge paths are not backed up
Fix the documentation string of the preference 'merge'.

Remove an old comment added when the the backup was (last) disabled.

Note that the backup of merge paths was documented but inactive at
initial commit b043e4c ("Initial import", 2005-02-07); it has been
activated at commit d166158 ("* Fixed some problems in the backup and
merging code, for better...", 2006-06-05) and disabled again at commit
2c898cc ("* Propagage changes from 2.27 branch.", 2008-01-01).