Main: catch and transform exception IllegalValue from the server into Fatal
This catches a discrepancy between server and client in the parsing of a
preference string and prints the error instead of the backtrace.

Even if such a discrepancy is probably best avoided by incompatible
version numbers, it could occur between development versions with the
same version, or because of a mistake or a local backport of new
features.