Pred: parse new negative pathspec of the form 'del <PATTERN>'
Note that a string associated with a pathspec is not overriden nor
deleted by a negative pathspec; a string given with a negative pathspec
is completely ignored.  At present this behaviour is fine as when a
pathspec preference is overriden by another (negative) preference
(ending in "not") the associated string is not used.  If this changes
the functions assoc and assoc_all may have to be adapted.

pred.mli: mention negative pathspec patterns.