Prefs: parse new directives "source" and "source?" (exactly matching "include")
The profile directives are like "include" and "include?" respectively
except that they never append a profile extension to the filename.

This is to prevent some unwanted infinite include loops (when one uses
both files with and without profile extensions and a file is
accidentally removed) as well as the inclusion of the wrong file (when a
file is accidentally created).