Pred: parse new assoc pathspec of the form 'assoc <PATTERN>'
The assoc patterns can record an associated string without having to set
the preference for the matching paths.  So a preference that requires an
associated string can be enabled independently someplace else.

The associated string is recorded but the pattern is completely ignored
for the compiled regex used to match a path in Pred.test.

pred.mli: update the documentation.