Pred: surround mapSeparator by spaces to split the line
This is to make it less likely to inadvertently specify an unwanted
separator in a pathspec.

The new separator is the one already used for the preference rootalias
(parsed in Update.root2stringOrAlias).

The code takes into account previously trimmed spaces (when the pattern
or the associated string is emtpy).