[build] Preliminary support for building with Dune.
Dune is a domain-specific build system that has become the standard
in the OCaml ecosystem.

This PR adds preliminary support for building Unison with Dune; the PR
itself is quite simple and doesn't modify the make-based system.

It has been tested in Linux and does work correctly. Among the new
features this PR provides are:

- automatic developer tools setup [merlin, etc...]
- Unison can now become supported in the OCaml platform,
- builds tested from 4.02.3 to 4.07.1,
- automatic generation of OPAM package / install files,
- massive linting of the codebase suggested [see warnings],
- compositional builds, easy library vendoring,
- hygienic builds.

I do believe this PR is important to bring Unison closer to the modern
OCaml toolchain. Additionally, some important platform changes are
upcoming, for example lablgtk2 is deprecated, and this PR will allow
to compose the Dune build with lablgtk3, greatly facilitating the
porting.

This PR opens up the question of what the supported OCaml version
are. 4.02.3 -- 4.07.1 seems like a sensible choice as of today.
Depending on the answer, we could then address:

- `-safe-string`: unison should be ported to -safe-string, however
  this usually places a lower-bound on OCaml 4.02.3.

- warnings: the codebase could benefit from the massive linting
  suggested by the compiler, however to make the warning system happy
  precise compiler version bounds have to be established.

TODO:

- Update CI
- Documentation ?
- OSX support:
  it shouldn't be too hard using the foreign sandboxing described at
  https://dune.readthedocs.io/en/latest/foreign-code.html#foreign-build-sandboxing
  unfortunately, I am not a Mac user.