Sync file modification time 0 (epoch)
Workaround to OCaml utimes behavior.

If atime and mtime arguments are both 0 then Unix.utimes will set
actual atime and mtime on the file to be the current timestamp, which
is not the desired result.

This is arguably not correct POSIX behavior, where NULL arguments will
set the actual times to current timestamp.

To sync the exact mtime value of 0, atime must be non-zero. Setting
atime to be different from zero by less than a nanosecond allows to
achieve the desired result and precise synchronization.

https://github.com/bcpierce00/unison/issues/223