Do no longer use Marshal in Fpcache
We introduce a new module, Umarshal, that provides basic
infrastructure for marshaling, along with some combinators, and use
that with the types that end up in fingerprint caches.

For each such type "typ", we define marshalling functions that are
wrapped in a record of type Umarshal.t, named "mtyp" or just "m" when
typ is "t".

The main substance of this commit is in Umarshal; all the other
changes are calls to the combinators defined in Umarshal to build
Umarshal.t values, and the switch from Marshal functions to Umarshal
ones in Fpcache.