Assume documentation strings are in UTF-8
This removes calls to transcodeDoc on documentation strings. The
transcodeDoc function transcodes from Windows Codepage 1252 to
Unicode, and assumes that the strings in strings.ml are in the former
encoding. Actually, they are in the encoding of the locale used to
generate strings.ml, which should be UTF-8 on most systems these days
(and was during the preparation of 2.51.3). Therefore, the call to
transcodeDoc is no longer needed.

Without this, the documentation (windows spawned by the "Help" menu)
contains corrupted characters.