uitext: selectAction: Fix tail call optimization
selectAction is a recursive function and so is the function calling it
to loop through all the updates. This is not a problem even with very
large number of updates since all involved functions are tail-recursive
and the compiler will optimize tail calls.

The optimization was not working because there was recursion in the
exception branch in selectAction. This commit removes the recursion from
the exception branch and enables the compiler to do correct tail call
optimization.