uimac: Fix Xcode build on OCaml versions >= 4.08
Since OCaml 4.08, Bigarray is included in Stdlib and -lbigarray is no
longer needed for linking. There is a compatibility shim provided, so
that linking with -lbigarray should not break with newer OCaml versions
but that does not seem to work and the builds fail on macOS. This commit
provides a workaround.