uimac: Type fixes to silence compiler warnings
NSInteger and NSUInteger are Apple types from Cocoa. They are defined
differently depending on the target architectures. On a 64-bit platform
they are 64-bit signed and unsigned integer, and they are 32-bit signed
and unsigned integer on 32-bit platforms.

Thus, we cast to long and use %ld (or unsigned long and %lu)--as is
commonly done with these types--assuming that NSInteger will always fit
in long. (We could cast to int64_t and use PRId64, assuming NSInteger
never becomes bigger than int64_t.)