Always commit archive updates
Unison uses files' mtime for fast update detection. This means that
mtime is stored in archive even if the value itself is not synchronized.

As mtime can have changed without file contents or other synchronized
properties having changed compared to the archive (or between replicas),
it is beneficial to record all updated mtimes in the archive also when
there were no updates synchronized. This will speed up update detection
the next time.

This logic is already included in the code but it misses the case where
there were no synchronizable updates detected at all. This change always
commits archive updates so that updated mtime values are recorded also
in case no updates where detected. It makes it consistent between UIs.