Add new directories as 'unchanged'
Currently, after synchronizing a newly added directory, it is always
treated as 'changed' during the next updates scanning, even if there
are no changes. This is not an issue but presents a small optimization
opportunity.

Synchronizing newly added directories as 'unchanged' (as they are just
synchronized, they are unchanged in terms of the archive) will allow
more efficient updates scanning the next time. This applies also to the
roots.

It is only a minor optimization as the effects are seen only at the
first scan after synchronizing a new directory and even then only if
the directory is actually unchanged since previous synchronization.