Change temp file prefixes to be OS-specific hidden files
The previous ".unison." temp file prefix was used on Windows,
where hidden files begin with "~$". This change detects if the
OS is Windows, and will make temp files properly hidden with
a "~$" prefix if so.

Unix file prefixes were updated to begin with ".~", which extends
support for certain backup applications that sync files beginning
with ".", but not ".~".