stasher: Fix creation of stashed directories
When creating a stashed directory in the backup directory then we must
check if such directory already exists or not. Previous code did this by
Os.exists which just checks for the absense of object. If the object was
present then the code just assumed it is a directory. This will cause
problems when the type of object changes (from a symlink to a directory,
for example).

The fix is to check if the name exists and is actually a directory. If
not then rotate backup versions until the new directory can be created.
The issue is only applicable with backuploaction "central" as this mode
creates a separate backup directory tree.