Try re-establishing broken connection at next scan
Unison relies on the remote connection being opened as part of the
profile initialization and then staying open for as long as Unison runs,
which in repeat mode can be an extended period of time. If the
connection breaks for whatever reason then it is treated as a fatal
error.

This patch catches the lost connection exception and the connection is
marked as closed. This allows to cleanly re-open the connection later.

Add function to check connection status, which as a side-effect triggers
throwing a lost connection exception if connection had been silently
broken, thereby detecting the closed state.

Add a function for UI to refresh connections by checking their status
and re-establishing them as necessary. This way the UI can rely on
remote connection being open at every scan even if it was temporarily
broken.

Just installing the roots again will also re-establish broken
connections, so even UIs that don't explicitly refresh connections will
benefit, as long as they reload the profile or re-install the roots.