Implement backwards-compatible feature negotiation
Features negotiation is an application-level mechanism to enable
extensibility in a backwards-compatible manner. This mechanism itself
is also a non-breaking change.

Any future changes that would otherwise be breaking changes can register
with this mechanism and be guarded to only work if enabled by feature
negotiation. If a client or a server of a different version does not
support the new code yet, then, instead of failing or refusing to connect
at all, it will be gracefully disabled (effectively acting as a version
downgrade for the connection).