fsmonitor: Report dir modify events on Windows
Current code does not report modification events on subdirectories.
Subdir creation, rename and deletion on the other hand are reported.
It is most likely done this way to prevent reporting "spurious"
modification events on the directory when in fact the list of files
in the directory had changed (which get reported by separate events).

This filtering will lose changes on the dir entity itself (think
attribute and ACL changes). The patch enables reporting modification
events on directories to make sure that these changes are not lost.