fsmonitor: Catch ENOENT during add_watch
While it is rare to occur, it is possible that a directory is deleted
after Unison has started scanning it but before fsmonitor has started
watching it. In this case the fsmonitor would fail with a fatal error.

This patch tries to be smart about this scenario and will fail only
if ENOENT is received on a root path. ENOENT on any sub-path is reported
as a normal change event. Unison will scan the path again and discover
that the directory is gone.