fsmonitor: Restore Windows 8.3 names processing
ebc5b832d45d156175659a5fc2fd1ea68b3b8fa5 incorrectly removed 8.3 short
names processing. MSDN includes this statement: "If there is both a
short and long name for the file, the function will return one of these
names, but it is unspecified which one."

This patch restores 8.3 short names processing. Instead of reverting the
previous patch, it implements a simpler mechanism which relies on the
assumption that 'watches' are always requested by long names. The
previous implementation relied on the same assumption.