Fix a row index error in the display loop
The status cell is set on the wrong row. It is not visible because
initially all status cells are empty. The real issue is considerable
performance degradation when large number of rows are displayed.
GtkCList probably has some internal exponential complexity relative to
row index; it is clearly seen when selecting/unselecting rows with large
indexes.