Change terminal input to read more than a byte
Currently only a single byte is read from the terminal when waiting for
user input. This is done both in line-buffered terminal mode and in the
'raw' non-line buffered terminal mode. This is not only sufficient, it
also has unexpected consequences in terms of interpreting user input. In
particular, since every single byte of the input is interpreted as a
separate user command then a multi-byte input effectively becomes a
sequence of commands.

This patch reads all the available input data -- the entire line in
line-buffered mode and all pending input (rather, up to 9 bytes) in
the 'raw' mode. The user input is no longer treated as a single byte.

This means that multi-byte encodings (for example, various UTF) will be
read properly, as an entire codepoint, not byte-per-byte. This also
means that ANSI escape sequences will be read properly, as an entire
sequence. The read input, even if it is multiple bytes, is treated as a
single user command.