Add transparent backwards compatibility for RPC
Use the old protocol in a non-intended (but safe) way to notify both the
server and the client that the other is capable of newer RPC version
negotiation. This is transparent to old clients.

Start connections in the backwards-compatible mode to enable 2.51 clients,
then upgrade if the client is known to be capable of it.