Listen on multiple sockets when needed
Previous code assumed there is always only one listening socket. That
one socket happened to be the first one that was successfully created
and bound. Usually, that would be IPv4 but it could just as well have
been IPv6.

This patch enables creating multiple listening sockets (for vast
majority of cases it is expected to be 2) and accepting clients on all
of these sockets. Like before, a single client at a time is served.