Make sending a preference to server conditional
Each preference can optionally specify a function that will be used to
determine if the preference should be sent to server. Typically, this
function will be of the form `fun () -> Features.enabled somefeature`,
which allows adding arbitrary preferences that will only be sent to the
server once feature negotiation has been concluded and a specific
feature agreed upon.