Store format-changing features in on-disk archive
Some features will impact the archive format (the structure). Those
features will be recorded in the archive file. This information is
required since the archive structure (schema) itself is not stored in
the file. The same features must exist and be known when loading the
archive to properly parse the file.