Add Bytes.get_*/set_* functions for OCaml < 4.08
Functions to read/write binary representations of numbers where added
to Stdlib Bytes and Buffer modules in OCaml 4.08. This patch adds a
compatibility module to expose a Bytes module with the same functions
available with OCaml < 4.08 (the primitives which the functions use
exist in earlier OCaml versions).

This patch can be reverted after the next release, when compatibility
with OCaml < 4.08 is no longer required.