fsmonitor: Remove concept of immediate events
Immediate events don't appear to serve any useful purpose. On the
contrary, they break correctness by reporting some events separately
from others and out of order.

This introduces a direct risk with "move" events that are reported by OS
as two separate events, delete + create. Delete events are currently
classified as immediate, thus reported separately from the create event.
This leads to Unison first synchronizing the delete, without the create,
which means that the other replica has completely lost the moved content.
If everything goes well, the next round of synchronization will catch up
and synchronize the create event, copying the content back to the other
replica. This is wasteful (not possible to use the local copy shortcut
at the other replica) even in the best case scenario. But if everything
does not go well, synchronization of the create may fail and leave the
other replica damaged from user's point of view.

Without the immedite event reporting this scenario is not possible
because Unison takes care to synchronize all the deletes at the very
end. In this case, if synchronizing the create event fails then the
other replica remains intact.

This patch removes the entire concept of immediate events, leaving all
events to be reported the same way and batched together.