Use Update.archive as Common.prevState
The primary intention of this change is to have more information
available about the previous archive in the updateItem. As a bonus,
this means less allocations and less memory consumption. This gain is
unfortunately lost in network transfer.

During network transfer, 0-8 bytes are added for each file change and up
to [file name length + ArchiveFile length (in range of roughly 57 to 137
bytes but can be considerably more)] bytes are added per each child,
recursively, for a directory deletion or with any change to any child
within the directory. When received over a network, in memory this data
structure requires for each child [name length + 202 to 282, and may be
considerably more] bytes. For example, deleting a largish directory of
100 000 files (regardless of depth of hierarchy) may easily add 10 MB to
the network transfer and require additional 25-35 MB of memory on the
receiving end. (Note that these sizes are not fixed, are counted _very_
roughly here, and can change from version to version.)