Don't assume stderr (or stdout) is always there
On Windows, it is possible for GUI applications to not have stdin,
stdout and stderr at all. Currently Unison (or any other non-Windows
software, I'd wager) does not account for this possibility and may
crash (I suspect the crash is actually in the OCaml libraries) upon
outputting to stderr (or stdout).