Don't open Windows console with the GUI
In Windows, Unison has always opened the Windows console even for the
GUI application. Historically, it has been required because there was no
pty support in Windows and also because the existence of working stdout
and stderr was taken for granted. The console does not need to be opened
by default and be visible at all times. It will be opened by Windows for
the ssh process, as required (or better yet, the new pty support will be
used). As for other stdout and stderr output, normal use of Unison GUI
should not produce any such output and does not require the console.