Open Windows console for debugging output
With Windows console no longer opened by default with the GUI
application, it must be opened on demand when debugging is requested
by the user. In Windows, for GUI applications without console, the
stdin, stdout and stderr don't exist, unless being redirected. They are
literally invalid, not just redirected to /dev/null or something.
Without opening the console, outputting to stderr (or stdout, whatever
is configured for debugging output) will crash the application.
(I suspect the crash is really in OCaml's libraries, but can't tell
where exactly.)

With this patch, the console is opened when running the GUI application,
any debug preference is set and either stdout or stderr is invalid.