GHA: Work around lablgtk build issues on Windows
lablgtk 2.18.12 builds on Windows hang in GHA CI if built with OCaml
versions < 4.10.0 (the actual issue is unknown, the GHA runners just time
out after a few hours). lablgtk 2.18.12 was released for compatibility
with upcoming OCaml 5, so the workaround is to just use previous lablgtk
versions on Windows with OCaml < 4.10.