Use format-changing features when loading archive
Since the archive structure (schema) is not stored in the on-disk
archive, any format-changing features are recorded in the file. Use this
information when loading the archive to make sure that the types match
those used when storing the archive. The marshal type functions (the
`m`) will use the information of enabled features.