Revert "Add a feature for OCaml 4.x hash function"
This reverts commit 56f253685a84d62f6a6bdc4bb7377f4a43c18b79.

This commit, while fully working and safe, was provided more as an
example of the "features" mechanism and was never intended as a
long-term solution.

There is no better replacement yet but the old compatibility mode works
well and is future-proof, so there is no real urgency in replacing it.