Update C code to latest FFI standards
Make sure C code adheres to latest FFI standards. This is required
mainly to be future-proof, but in some cases this may also be relevant
to ensure memory-safety with upcoming multicore OCaml (>= 5.00).

In particular, the patch includes the following updates:

 - Use correct names (as if CAML_NAME_SPACE was defined); the old
   symbols will no longer be available starting with OCaml 5.00
 - Always use CAMLparam, CAMLlocal and CAMLreturn, even if values are
   known to be unboxed and even if there are no allocations. This makes
   for a consistent approach and does not rely on certain values having
   to be unboxed. The consistent approach is extra important starting
   with OCaml 5.00 when previous single-threaded assumptions may no
   longer hold.
 - Add an explicit `unit` parameter for functions that have no other
   parameters.
 - Remove duplication of functions where the only difference is that of
   accessing a value as either string or bytes.
 - Remove prototype (re-)definitions that are already available via
   OCaml headers.

The patch does not replace the usage of `Field()` with `Store_field()`.
While the former may in some cases not be safe for multi-threaded use,
the issues may have been resolved by the time OCaml 5.00 is released.
Further, as used in this code, `Field()` is not considered a risk.