Capture stderr of the ssh process, take two
Previously, for the GTK GUI stderr output from the ssh process was
directed to Unix.stderr. This is limiting because of two reasons:
a GUI user may never see the stderr output; and in Windows the stderr
may not even be a valid fd (a GUI-only application).

Commit 763c5d85298e7e9a03a920439cfa513647adb3dc redirected stderr of the
ssh process to the controlling terminal. The terminal's output was
already displayed in the GUI (for password prompts) and now all stderr
output from ssh was also being displayed to the user.

While ssh also forwards the stderr from the remote host, capturing and
displaying stderr of the ssh process in the GUI is primarily intended to
bring to user's attention errors from ssh itself, such as connection and
authentication errors.

Due to primitive code for displaying password prompts from ssh, and
even though ssh's stdout and stderr both were captured from the same
controlling terminal, the output from stderr was displayed
inconsistently in the GUI. A hack trying to fix this was included in
commit f13eb1d7222645ad4d549cbaa0de67e5d4feee4a.

All of the above was still not sufficient for even a somewhat-working
interface towards users. Moreover, blindly trying to capture all stderr
output from ssh and display it in GUI could result in GUI hangs because
any output from the remote unison process would be attempted to be
displayed as a password prompt, even if in the middle of syncing (for
example, debug output of the server is produced on stderr).

This patch is an attempt at getting this right (but far from ideal).

 - stderr output is displayed in GUI only up to detecting an established
   connection. This separates error output from the ssh process and from
   the remote unison process.

 - After the connection has been established, stderr output is forwarded
   to local stderr. This is expected for remote debug logs.

 - The logic for reading from controlling terminal and prompting in GUI
   and writing back the response has been separated into two Lwt threads
   as an attempt (but it is not a definitive solution) to present output
   from ssh more correctly to the user. It works by creating a separate
   buffer and displaying its contents in GUI after a small time delay,
   allowing the buffer to be filled with multiple writes from the ssh
   process.