Extend the backwards compatibility foundation
This commit is a logical continuation to the work started in commit
90dd589dc7cfa99b0eaa73a5cc1d202cfe0fc1ee. Without actually changing the
types, it provides the foundation so that they can be changed while
keeping compatibility with versions <= 2.51.5.

The types prepared like this are [Fileinfo.stamp] and [Props.t]
(continuing from the earlier commit) and other types which include any
of these two or [archive] (which itself was prepared in the earlier
commit).

A bug from commit aa29f5d6b2b0b2c26c8d5d0cca56ff1167ebac40 is fixed.