Start NEWS.md in the sources, for 2.52.0
The manual build process has in practice had some friction, and the
sources don't currently meet the widespread expectation of having NEWS
at top level.  This commit starts NEWS.md for 2.52.0, and notes in the
manual that news is now where most projects have it.

It is TBD whether to migrate the 2.51.x news items to NEWS.md or not.