Enhance the bind preference of socket server
Rename "host" command-line argument to "listen" (keep old name as
alias) and make it repeatable. Now the user can supply this argument
multiple times to bind and listen on several different addresses.
Clarify this in the pref documentation.

As a separate fix, make sure that aliases work properly when preferences
are sent from client to server (fixes compatibility issue).