Make socket mode work with Unix domain sockets
Enhance the current socket mode so that it works also with Unix domain
sockets. In particular, the 'socket' command-line argument is enhanced
to receive either a filesystem path or a port number. The roots syntax
is enhanced respectively so that a filesystem path for a local socket
can be provided in place of a host name.