uigtk: Don't force quit the GUI with fatal errors
Add a Close button to the fatal error message dialog. Clean up and
structure the code to either leave the GUI in the neutral state and
allow the user to try again or change the profile; or force a profile
change if still in the early init process.

Clean up global state and guard the code against continuing after an
interrupted action.

Hint:
  the following git options may help when reviewing

  -w --color-moved=blocks --color-moved-ws=ignore-all-space