Ignore profile problems when scanning profile list
Previously, an error during initial scanning the profiles would force
the user interface to quit without user being able to rectify the
situation (indeed, nor even understand what is going on). This would
happen at an error in _any_ profile, not just the one the user had
requested.

Since this scanning is only done to build a list of profiles for the
profile manager, we can safely ignore any errors at this stage (while
still notifying the user).