uigtk2: build fast-profiles menu dynamically
This is to better organise code and not force scanning the profiles
before the GUI is built (any error/warning during the scanning would
then go lost or force the exit before the GUI is even built).

Later, should be updated to update the menu every time profile manager
was opened (because profiles may have been added/removed/changed).